Refactored code so that classes are seperate
