+++ src/classes/Block.js
@@ -4,10 +4,18 @@ export class Block {
     
     constructor(index, timestamp, data, previousHash = '') {
         this.index = index;
-        this.timestamp = timestamp;
+        this.timestamp = this.setDate();
         this.data = data;
         this.previousHash = previousHash;
         this.hash = this.calculateHash;
+    }
+
+    setDate() {
+        var date = new Date();
+        var currentDateTime = date.getDate() + "/" + date.getMonth() + "/" + date.getFullYear()
+                            + " - " + date.getHours() + ":" + date.getMinutes() + ":" + date.getSeconds();
+        
+        return currentDateTime;
     }
 
     calculateHash() {
